project/requirements.txt
Django>=4.0
djangorestframework
transformers
fuzzywuzzy
python-docx
PyPDF2
